Refactored process management API endpoints for listing, status, termination, pause, resume, and dashboard
